Handle get_next_line if exhasusted, and force encoding to fix encoding issue
Not really sure if this is an appropriate long term solution, but side steps the
issue I was running into for now. Shoudl dig into seeing if crawdad can be
modified to work in UTF-8 or if the ASCII 8bit encoding is coming from somewhere
else that can be modified.